updated the client authentication to be agnostic to what component is calling it, also added client login
